requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
build==1.4.4
certifi==2026.4.22
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.3.3
coloredlogs==15.0.1
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.1
fastapi==0.136.1
filelock==3.29.0
flatbuffers==25.12.19
fsspec==2026.3.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.12.2
humanfriendly==10.0
idna==3.13
importlib_metadata==8.7.1
importlib_resources==7.1.0
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.33.0
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orjson==3.11.8
overrides==7.7.0
packaging==26.2
protobuf==6.33.6
pybase64==1.4.3
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.23.1
tomli==2.4.1
tqdm==4.67.3
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
zipp==3.23.1
